backend/requirements.txt
django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
django-celery-beat==2.6.0
django-redis==5.4.0
django-storages==1.14.4
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.0
channels==4.1.0
channels-redis==4.2.0
uvicorn[standard]==0.30.0
httpx==0.27.0
gunicorn==22.0.0
whitenoise==6.7.0
Pillow==10.4.0
weasyprint==62.3
python-dotenv==1.0.1
dj-database-url==2.2.0
openai==1.101.0
markdown==3.7
bleach==6.1.0
